AutoGen/requirements.txt
python-dotenv
ag2[openai]
pandas
numpy
matplotlib
yfinance
requests